% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/operation-adjust.R
\docType{class}
\name{Adjust-class}
\alias{Adjust-class}
\alias{adjust}
\title{Adjust layer params and aesthetics (Layer Operation)}
\usage{
adjust(x, mapping = aes(), ...)
}
\arguments{
\item{x}{One of:
\itemize{
\item A \link{Layer}-like object: applies this operation to the layer.
\item A missing argument: creates an \link{Operation}
\item Anything else: creates an \link{Operation}, passing \code{x} along to the
\code{mapping} argument
}}

\item{mapping}{An aesthetic created using \code{aes()}. Mappings provided here
will overwrite mappings in \link{Layer}s when this \link{Operation} is applied to
them.}

\item{...}{\link{Layer} parameters. Params provided here will overwrite
params in \link{Layer}s when this \link{Operation} is applied to them.}
}
\value{
A \link{Layer}-like object (if \code{x} is a \link{Layer}) or an \link{Operation} (if not).
}
\description{
A layer \link{Operation} for adjusting the params and aesthetic mappings of
a \link{Layer}.
}
\examples{

library(ggplot2)

# here we use adjust() with nop() ( + 1) to create a copy of
# the stat_smooth layer, putting a white outline around it.
set.seed(1234)
k = 1000
data.frame(
  x = seq(1, 10, length.out = k),
  y = rnorm(k, seq(1, 2, length.out = k) + c(0, 0.5)),
  g = c("a", "b")
) |>
  ggplot(aes(x, y, color = g)) +
  geom_point() +
  stat_smooth(method = lm, formula = y ~ x, size = 1.5, se = FALSE) *
    (adjust(aes(group = g), color = "white", size = 4) + 1) +
  scale_color_brewer(palette = "Dark2")

# (note this could also be done with copy_under())

}
\seealso{
\link{Operation} for a description of \link{Layer} operations.

Other Layer operations: 
\code{\link{Nop-class}},
\code{\link{copy}}
}
\concept{Layer operations}
