% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blend.R
\name{blend}
\alias{blend}
\title{Blend ggplot2 layers}
\usage{
blend(x, blend = "over", alpha = 1, ...)
}
\arguments{
\item{x}{A \code{ggplot2::Layer}, such as a \code{geom} or \code{stat}, or a list
of layers.}

\item{blend}{The blend mode to use. The default mode, \code{"over"}, corresponds to
the "usual" blend mode of drawing objects on top of each other.
The list of supported blend modes depends on your graphical device
(see Murrell 2021), and are listed in \code{dev.capabilities()$compositing}.
Blend modes can include: \code{"clear"},
\code{"source"}, \code{"over"}, \code{"in"}, \code{"out"}, \code{"atop"}, \code{"dest"}, \code{"dest.over"},
\code{"dest.in"}, \code{"dest.out"}, \code{"dest.atop"}, \code{"xor"}, \code{"add"}, \code{"saturate"},
\code{"multiply"}, \code{"screen"}, \code{"overlay"}, \code{"darken"}, \code{"lighten"},
\code{"color.dodge"}, \code{"color.burn"}, \code{"hard.light"}, \code{"soft.light"},
\code{"difference"}, and \code{"exclusion"}

Blend modes like \code{"multiply"} and \code{"screen"} are particularly useful as they
are \emph{commutative}: the result is the same whichever order they are applied in.}

\item{alpha}{A numeric between \code{0} and \code{1} (inclusive). The opacity of a
transparency mask applied to objects prior to blending.}

\item{...}{Additional arguments (currently unused).}
}
\value{
An object that can be added to a \code{ggplot()} object: if the input is a \code{ggplot2::Layer},
the result is a \code{ggplot2::Layer}; if the input is a \code{list} of \code{ggplot2::Layer}s,
the output is a \code{list} of \code{ggplot2::Layer}s.
}
\description{
Blend objects within a single layer (geom) or across multiple layers (geoms)
using graphical blending modes, such as \code{"multiply"}, \code{"overlay"}, etc. Uses
the built-in compositing support in graphical devices added in R 4.2.
}
\details{
If \code{x} is a single layer / geometry and the \code{blend_group} aesthetic \emph{is not} set, every
graphical object (\code{grob()}) output by the geometry will be blended together
using the \code{blend} blend mode. If \code{alpha != 1}, a transparency mask with the
provided alpha level will be applied to each grob before blending.

If \code{x} is a single layer / geometry and the \code{blend_group} aesthetic \emph{is} set,
the geometry will be rendered for each subset of the data defined by the
\code{blend_group} aesthetic, a transparency mask with the provided \code{alpha} level
will be applied to each resulting group as a whole (if \code{alpha != 1}), then these groups
will be blended together using the \code{blend} blend mode.

If \code{x} is a list of layers / geometries, those layers will be rendered
separately, a transparency mask with the provided \code{alpha} level
will be applied to each layer as a whole (if \code{alpha != 1}), then these layers
will be blended together using the \code{blend} blend mode.
}
\references{
Murrell, Paul (2021):
\href{https://www.stat.auckland.ac.nz/~paul/Reports/GraphicsEngine/groups/groups.html}{Groups, Compositing Operators, and Affine Transformations in R Graphics}.
The University of Auckland. Report.
\doi{10.17608/k6.auckland.17009120.v1}.
}
\seealso{
\code{\link[=stack_blends]{stack_blends()}}
}
