% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggblend-package.R
\docType{package}
\name{ggblend-package}
\alias{ggblend-package}
\alias{ggblend}
\title{Blending and compositing for ggplot2}
\description{
\pkg{ggblend} is an R package that adds support for R 4.2 blend modes
(e.g. \code{"multiply"}, \code{"overlay"}, etc) to \pkg{ggplot2}.
}
\details{
Support for blending layers (e.g. \code{geom}s) in \pkg{ggplot2}. See \code{blend()}.
}
\section{Package options}{


The following global options can be set using \code{options()} to modify the
behavior of \pkg{ggblend}:
\itemize{
\item \code{"ggblend.check_blend"}: If \code{TRUE} (default), \code{blend()} will warn if
you attempt to use a blend mode not supported by the current graphics
device, as reported by \code{dev.capabilities()$compositing}. Since this check
can be unreliable on some devices (they will report not support a blend
mode that they do support), you can disable this warning by setting this
option to \code{FALSE}.
}
}

