% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/operation-.R,
%   R/operation-adjust.R, R/operation-nop.R, R/operation-product.R
\docType{class}
\name{Operation-class}
\alias{Operation-class}
\alias{Operation}
\alias{show,Operation-method}
\alias{format,Operation-method}
\alias{format,Adjust-method}
\alias{format,Nop-method}
\alias{format,OperationProduct-method}
\title{Layer operations}
\usage{
\S4method{show}{Operation}(object)

\S4method{format}{Operation}(x, ...)

\S4method{format}{Adjust}(x, ...)

\S4method{format}{Nop}(x, ...)

\S4method{format}{OperationProduct}(x, ...)
}
\arguments{
\item{x, object}{An \link{Operation}.}

\item{...}{Further arguments passed to other methods.}
}
\description{
\link{Operation}s are composable transformations that can be applied to \pkg{ggplot2}
\link{Layer}s or \link{Layer}-like objects (e.g. lists of \link{Layer}s; see the \link{Layer}
documentation page for a description of valid \link{Layer}-like objects).
}
\details{
\link{Operation}s can be composed using the \code{+} and \code{*} operators (see \link{OperationSum}
and \link{OperationProduct}). Addition and multiplication of \link{Operation}s and \link{Layer}s
obeys the distributive law.

\link{Operation}s can be applied to \link{Layer}s using \code{*} or \verb{|>}, with slightly
different results:
\itemize{
\item Using \code{*}, application of \link{Operation}s to a list of \link{Layer}s \emph{is} distributive. For example,
\code{list(geom_line(), geom_point()) * blend("multiply")} is
equivalent to \code{list(geom_line() * blend("multiply"), geom_point() * blend("multiply"))};
i.e. it multiply-blends the contents of the two layers individually.
\item Using \verb{|>}, application of \link{Operation}s to a list of \link{Layer}s is \emph{not}
distributive (unless the only reasonable interpretation of applying the
transformation is necessarily distributive; e.g. \code{adjust()}). For example,
\code{list(geom_line(), geom_point()) |> blend("multiply")} would multiply-blend
both layers together, rather than multiply-blending the contents of the
two layers individually.
}
}
\section{Functions}{
\itemize{
\item \code{show,Operation-method}: Print an \link{Operation}.

\item \code{format,Operation-method}: Format an \link{Operation} for printing.
}}

