% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/operation-.R,
%   R/operation-adjust.R, R/operation-product.R
\docType{class}
\name{OperationProduct-class}
\alias{OperationProduct-class}
\alias{*,Operation,ANY-method}
\alias{*,ANY,Operation-method}
\alias{*,Adjust,Adjust-method}
\alias{OperationProduct}
\alias{prod,Operation-method}
\alias{*,Operation,Operation-method}
\alias{*,numeric,Operation-method}
\alias{*,Operation,numeric-method}
\alias{*,Operation,OperationSum-method}
\alias{*,OperationSum,Operation-method}
\alias{*,OperationSum,OperationSum-method}
\title{Layer operation multiplication and application}
\usage{
\S4method{*}{Operation,ANY}(e1, e2)

\S4method{*}{ANY,Operation}(e1, e2)

\S4method{*}{Adjust,Adjust}(e1, e2)

\S4method{prod}{Operation}(x, ..., na.rm = FALSE)

\S4method{*}{Operation,Operation}(e1, e2)

\S4method{*}{numeric,Operation}(e1, e2)

\S4method{*}{Operation,numeric}(e1, e2)

\S4method{*}{Operation,OperationSum}(e1, e2)

\S4method{*}{OperationSum,Operation}(e1, e2)

\S4method{*}{OperationSum,OperationSum}(e1, e2)
}
\arguments{
\item{e1}{an \link{Operation}, \link{Layer}, or \code{\link[=numeric]{numeric()}}}

\item{e2}{an \link{Operation}, \link{Layer}, or \code{\link[=numeric]{numeric()}}}

\item{x, ...}{\link{Operation}s}

\item{na.rm}{ignored}
}
\description{
\link{Operation}s can be multiplied together to form chains of operations, which
when multiplied by (applied to) \link{Layer}s, return modified \link{Layer}s.
}
\details{
Multiplication of \pkg{ggblend} \link{Operation}s depends on the types of
objects being multiplied:
\itemize{
\item If you multiply an \link{Operation} with an \link{Operation}, they are merged into
a single \link{Operation} that applies each \link{Operation} in sequence.
\item If you multiply an \link{Operation} with a \link{Layer}, that operation is applied
to the layer, returning a new \link{Layer}.
\item If you multiply an \link{Operation} by a \code{\link[=numeric]{numeric()}} \emph{n}, a new \link{Operation} that
repeats the input \link{Operation} is \emph{n} times is returned.
}
}
\examples{

library(ggplot2)

# multiplying operations by numerics repeats them...
adjust(color = "red") * 2

# multiplying operations together chains (or merges) them
adjust(color = "red") * adjust(size = 2)

# multiplication obeys the distributive law
op = (adjust(aes(y = 11 -x), color = "skyblue") + 1) * (adjust(color = "white", size = 4) + 1)
op

# multiplication by a geom returns a modified version of that geom
data.frame(x = 1:10) |>
  ggplot(aes(x = x, y = x)) +
  geom_line(size = 2) * op
}
