% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/operation-sum.R
\docType{class}
\name{OperationSum-class}
\alias{OperationSum-class}
\alias{OperationSum}
\alias{sum,Operation-method}
\alias{+,Operation,Operation-method}
\alias{+,Operation,numeric-method}
\alias{+,numeric,Operation-method}
\alias{format,OperationSum-method}
\title{Layer operation addition}
\usage{
\S4method{sum}{Operation}(x, ..., na.rm = FALSE)

\S4method{+}{Operation,Operation}(e1, e2)

\S4method{+}{Operation,numeric}(e1, e2)

\S4method{+}{numeric,Operation}(e1, e2)

\S4method{format}{OperationSum}(x, ...)
}
\arguments{
\item{x, ...}{\link{Operation}s}

\item{na.rm}{ignored}

\item{e1}{an \link{Operation} or \code{\link[=numeric]{numeric()}}}

\item{e2}{an \link{Operation} or \code{\link[=numeric]{numeric()}}}
}
\description{
\link{Operation}s can be added together to form stacks of operations, which
when multiplied by (applied to) \link{Layer}s, those \link{Layer}s are distributed
over the \link{Operation}s (i.e. copied).
}
\details{
Addition of \pkg{ggblend} \link{Operation}s depends on the types of
objects being summed:
\itemize{
\item If you add an \link{Operation} to an \link{Operation}, they are merged into
a single \link{Operation} that copies input \link{Layer}s, one for each \link{Operation}.
\item If you add an \link{Operation} to a \code{\link[=numeric]{numeric()}} \emph{n}, it is equivalent to
adding \code{*} \code{\link[=nop]{nop()}}s to that \link{Operation}.
}
}
\examples{

library(ggplot2)

# adding operations together creates a sum of operations
adjust(color = "red") + adjust(size = 2)

# addition and multiplication obey the distributive law
op = (adjust(aes(y = 11 -x), color = "skyblue") + 1) * (adjust(color = "white", size = 4) + 1)
op

# multiplication by a geom returns a modified version of that geom,
# distributed over the sum of the operations
data.frame(x = 1:10) |>
  ggplot(aes(x = x, y = x)) +
  geom_line(size = 2) * op
}
