% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation-partition.R
\name{partition}
\alias{partition}
\title{Partition a layer into subgroups (Layer operation)}
\usage{
partition(x, partition)
}
\arguments{
\item{x}{One of:
\itemize{
\item A \link{layer-like} object: applies this operation to the layer.
\item A missing argument: creates an \link{operation}
\item Anything else: creates an \link{operation}, passing \code{x} along to the
\code{partition} argument
}}

\item{partition}{One of:
\itemize{
\item A list of quosures, such as returned by \code{\link[=vars]{vars()}}, giving a (possibly multi-)
column expression for the \code{partition} aesthetic. These expressions are
combined using \code{\link[=interaction]{interaction()}} to be passed on to \code{aes(partition = ...)}
\item A one-sided formula, giving a single-column expression for the \code{partition}
aesthetic, which is passed on to \code{aes_(partition = ...)}.
}}
}
\value{
A \link{layer-like} object (if \code{x} is \link{layer-like}) or an \link{operation} (if not).
}
\description{
A layer \link{operation} for adding a \code{partition} aesthetic to a \link{layer}.
}
\details{
This is a shortcut for setting the \code{partition} aesthetic of a \link{layer}.
\itemize{
\item \code{partition(~ XXX)} is roughly equivalent to \code{adjust(aes(partition = XXX))}
\item \code{partition(vars(X, Y, ...))} is roughly equivalent to \code{adjust(aes(partition = interaction(X, Y, ...)))}
}

When a \link{layer} with a \code{partition} aesthetic is used by the following
\link{operation}s, the effects of the operations are applied across groups:
\itemize{
\item \code{\link[=blend]{blend()}}: Blends graphical objects within the subgroups defined by the
partition together using normal (\code{"over"}) blending before applying its
blend between subgroups.
}
}
\examples{
library(ggplot2)

# TODO

}
\seealso{
\link{operation} for a description of layer operations.

Other layer operations: 
\code{\link{adjust}},
\code{\link{blend}},
\code{\link{copy}},
\code{\link{nop}}
}
\concept{layer operations}
