% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggblend-class.R
\name{pow-ggblend}
\alias{pow-ggblend}
\alias{^.ggblend}
\alias{^,ggblend,ggblend-method}
\title{Compose blends into a stack}
\usage{
\S4method{^}{ggblend,ggblend}(e1, e2)
}
\arguments{
\item{e1}{A \code{"ggblend"} object as returned by \code{\link[=blend]{blend()}}, \code{\link[=stack_blends]{stack_blends()}}, etc.}

\item{e2}{A \code{"ggblend"} object as returned by \code{\link[=blend]{blend()}}, \code{\link[=stack_blends]{stack_blends()}}, etc.}
}
\description{
Operator for stacking blends: \code{blend(..1) ^ blend(..2)} is equivalent to \code{stack_blends(..1, ..2)}.
}
\details{
An algebraic alternative syntax to \code{\link[=stack_blends]{stack_blends()}}.
}
\examples{
library(ggplot2)

# create two versions of a dataset, where draw order can affect output
set.seed(1234)
df_a = data.frame(x = rnorm(500, 0), y = rnorm(500, 1), set = "a")
df_b = data.frame(x = rnorm(500, 1), y = rnorm(500, 2), set = "b")
df_ab = rbind(df_a, df_b) |>
  transform(order = "draw a then b")
df_ba = rbind(df_b, df_a) |>
  transform(order = "draw b then a")
df = rbind(df_ab, df_ba)

# it can be useful to stack multiple versions of the same layer on top of
# each other to create a desired effect: here, `geom_point()` is used
# twice to lighten the multiply effect (otherwise points become very dark)
df |>
  ggplot(aes(x, y, color = set)) +
  geom_point(size = 3) * blend("lighten") +
  geom_point(size = 3) * blend("multiply", alpha = 0.65) +
  scale_color_brewer(palette = "Set2") +
  facet_grid(~ order)

# we can reduce code duplication by using `^`
df |>
  ggplot(aes(x, y, color = set)) +
  geom_point(size = 3) * blend("lighten") ^ blend("multiply", alpha = 0.65) +
  scale_color_brewer(palette = "Set2") +
  facet_grid(~ order)

}
\seealso{
Other blending functions and operators: 
\code{\link{blend}},
\code{\link{stack_blends}()},
\code{\link{times-ggblend}}
}
\concept{blending functions and operators}
