% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggblend-class.R, R/ggblend_stack-class.R
\name{print.ggblend}
\alias{print.ggblend}
\alias{show,ggblend-method}
\alias{print.ggblend_stack}
\alias{show,ggblend_stack-method}
\title{Print a ggblend object}
\usage{
\method{print}{ggblend}(x, ...)

\S4method{show}{ggblend}(object)

\method{print}{ggblend_stack}(x, ...)

\S4method{show}{ggblend_stack}(object)
}
\arguments{
\item{x, object}{A \code{"ggblend"} object.}

\item{...}{Additional arguments passed to other methods}
}
\value{
\code{x} / \code{object}, invisibly.
}
\description{
Displays a \code{"ggblend"} object.
}
\examples{

print(blend("multiply", alpha = 0.5))

}
