% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation-copy.R
\name{copy}
\alias{copy}
\alias{copy_over}
\alias{copy_under}
\title{Copy layers then adjust params and aesthetics (Layer operation)}
\usage{
copy_over(mapping = aes(), ...)

copy_under(mapping = aes(), ...)
}
\arguments{
\item{mapping}{An aesthetic created using \code{aes()}. Mappings provided here
will overwrite mappings in \link{Layer}s when this \link{operation} is applied to
them.}

\item{...}{\link{Layer} parameters. Params provided here will overwrite
params in \link{Layer}s when this \link{operation} is applied to them.}
}
\value{
A \link{Layer}-like object (if \code{x} is a \link{Layer}) or an \link{operation} (if not).
}
\description{
A layer \link{operation} for copying and then adjusting the params and aesthetic
mappings of a \link{Layer}.
}
\details{
These are shortcuts for duplicating a layer and then applying \code{\link[=adjust]{adjust()}}.
Specifically:
\itemize{
\item \code{copy_over(...)} is equivalent to \code{1 + adjust(...)}
\item \code{copy_under(...)} is equivalent to \code{adjust(...) + 1}
}
}
\examples{
library(ggplot2)

# here we use copy_under() to create a copy of
# the stat_smooth layer, putting a white outline around it.
set.seed(1234)
k = 1000
data.frame(
  x = seq(1, 10, length.out = k),
  y = rnorm(k, seq(1, 2, length.out = k) + c(0, 0.5)),
  g = c("a", "b")
) |>
  ggplot(aes(x, y, color = g)) +
  geom_point() +
  stat_smooth(method = lm, formula = y ~ x, size = 1.5, se = FALSE) *
    copy_under(aes(group = g), color = "white", size = 4) +
  scale_color_brewer(palette = "Dark2")

}
\seealso{
\link{operation} for a description of \link{Layer} operations.

Other Layer operations: 
\code{\link{adjust}},
\code{\link{nop}}
}
\concept{Layer operations}
