% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-operation-.R, R/operation-.R,
%   R/operation-adjust.R, R/operation-blend.R, R/operation-composition.R,
%   R/operation-nop.R, R/operation-product.R
\docType{class}
\name{operation-class}
\alias{operation-class}
\alias{operation}
\alias{show,operation-method}
\alias{format,operation-method}
\alias{format,adjust-method}
\alias{format,blend-method}
\alias{format,operation_composition-method}
\alias{format,nop-method}
\alias{format,operation_product-method}
\title{Layer operations}
\usage{
\S4method{show}{operation}(object)

\S4method{format}{operation}(x, ...)

\S4method{format}{adjust}(x, ...)

\S4method{format}{blend}(x, ...)

\S4method{format}{operation_composition}(x, ...)

\S4method{format}{nop}(x, ...)

\S4method{format}{operation_product}(x, ...)
}
\arguments{
\item{x, object}{An \link{operation}.}

\item{...}{Further arguments passed to other methods.}
}
\description{
\link{operation}s are composable transformations that can be applied to \pkg{ggplot2}
\link{Layer}s or \link{Layer}-like objects (e.g. lists of \link{Layer}s; see the \link{Layer}
documentation page for a description of valid \link{Layer}-like objects).
}
\details{
\link{operation}s can be composed using the \code{+} and \code{*} operators (see \link{operation_sum}
and \link{operation_product}). Addition and multiplication of \link{operation}s and \link{Layer}s
obeys the distributive law.

\link{operation}s can be applied to \link{Layer}s using \code{*} or \verb{|>}, with slightly
different results:
\itemize{
\item Using \code{*}, application of \link{operation}s to a list of \link{Layer}s \emph{is} distributive. For example,
\code{list(geom_line(), geom_point()) * blend("multiply")} is
equivalent to \code{list(geom_line() * blend("multiply"), geom_point() * blend("multiply"))};
i.e. it multiply-blends the contents of the two layers individually.
\item Using \verb{|>}, application of \link{operation}s to a list of \link{Layer}s is \emph{not}
distributive (unless the only reasonable interpretation of applying the
transformation is necessarily distributive; e.g. \code{adjust()}). For example,
\code{list(geom_line(), geom_point()) |> blend("multiply")} would multiply-blend
both layers together, rather than multiply-blending the contents of the
two layers individually.
}
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print an \link{operation}.

\item \code{format}: Format an \link{operation} for printing.
}}

