% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operation-partition.R
\name{partition}
\alias{partition}
\title{Partition a layer into subgroups (Layer operation)}
\arguments{
\item{x}{One of:
\itemize{
\item A \link{Layer}-like object: applies this operation to the layer.
\item A missing argument: creates an \link{operation}
\item Anything else: creates an \link{operation}, passing \code{x} along to the
\code{partition} argument
}}

\item{partition}{A quoted column expression for the \code{partition}
aesthetic, to be passed on to \code{aes_()}; e.g. a \link{formula},
\link{quosure}, or output from \code{\link[=vars]{vars()}}.}
}
\value{
A \link{Layer}-like object (if \code{x} is a \link{Layer}) or an \link{operation} (if not).
}
\description{
A layer \link{operation} for adding a \code{partition} aesthetic to a \link{layer}.
}
\details{
This is a shortcut for setting the \code{partition} aesthetic of a \link{layer}.
\code{partition(~ XXX)} is equivalent to \code{adjust(aes(partition = XXX))}.

When a \link{layer} with a \code{partition} aesthetic is used by the following
\link{operation}s, the effects of the operations are applied to each subgroup
defined by the partition: \code{\link[=blend]{blend()}}.
}
\examples{
library(ggplot2)

# TODO

}
\seealso{
\link{operation} for a description of \link{Layer} operations.

Other Layer operations: 
\code{\link{adjust}},
\code{\link{blend}},
\code{\link{copy}},
\code{\link{nop}}
}
\concept{Layer operations}
