% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_blends.R
\name{stack_blends}
\alias{stack_blends}
\alias{stack_blends.Layer}
\alias{stack_blends.list}
\alias{stack_blends.character}
\alias{stack_blends.ggblend}
\title{Stack multiple blends of the same object on top of each other}
\usage{
stack_blends(x, ...)

\method{stack_blends}{Layer}(x, ...)

\method{stack_blends}{list}(x, ...)

\method{stack_blends}{character}(x, ...)

\method{stack_blends}{ggblend}(x, ...)
}
\arguments{
\item{x}{One of:
\itemize{
\item A \code{ggplot2::Layer}, such as a \code{geom} or \code{stat}, or a list of layers.
\item A string (character vector of length 1) giving the name of a blend,
which is assumed to be the first blend in the sequence.
\item A \code{"ggblend"} object as returned by \code{\link[=blend]{blend()}}, \code{\link[=stack_blends]{stack_blends()}}, etc.
}

If a layer is provided, this stack of blends is applied to it.
If a string or \code{"ggblend"} object is provided, a new
\code{"ggblend"} object is returned which can be applied using the
\code{\link{*.ggblend}} and \code{\link{^.ggblend}} operators.}

\item{...}{A sequence of lists of arguments to pass to \code{\link[=blend]{blend()}}, each
representing a blend applied to a copy of \code{x}. For example, a single string
(like \code{"multiply"}) would result in a \code{blend(x, "multiply")} layer being
created; a list (like \code{list("multiply", alpha = 0.5)}) would result
in a \code{blend(x, "multiply", alpha = 0.5)} layer being created.}
}
\value{
One of:
\itemize{
\item If the input is a \code{ggplot2::Layer} or list, the output is a list of \code{ggplot2::Layer}s
that can be added to a \code{\link[=ggplot]{ggplot()}} object.
\item If the input is a string or a \code{"ggblend"}, the output is a \code{"ggblend"} object,
which can be combined with other \code{"ggblend"} objects using \code{\link{^.ggblend}} or
applied to a ggplot layer using \code{\link{*.ggblend}}.
}
}
\description{
This is a shortcut for duplicating a layer multiple times, applying a
different blend to each layer, then stacking the results on top of each
other. This can be useful when a single blend in isolation does not achieve
the desired effect (e.g. when dealing with overlapping points).
}
\examples{
library(ggplot2)

# create two versions of a dataset, where draw order can affect output
set.seed(1234)
df_a = data.frame(x = rnorm(500, 0), y = rnorm(500, 1), set = "a")
df_b = data.frame(x = rnorm(500, 1), y = rnorm(500, 2), set = "b")
df_ab = rbind(df_a, df_b) |>
  transform(order = "draw a then b")
df_ba = rbind(df_b, df_a) |>
  transform(order = "draw b then a")
df = rbind(df_ab, df_ba)

# it can be useful to stack multiple versions of the same layer on top of
# each other to create a desired effect: here, `geom_point()` is used
# twice to lighten the multiply effect (otherwise points become very dark)
df |>
  ggplot(aes(x, y, color = set)) +
  geom_point(size = 3) |> blend("lighten") +
  geom_point(size = 3) |> blend("multiply", alpha = 0.65) +
  scale_color_brewer(palette = "Set2") +
  facet_grid(~ order)

# we can reduce code duplication by using stack_blends()
df |>
  ggplot(aes(x, y, color = set)) +
  geom_point(size = 3) |> stack_blends("lighten", list("multiply", alpha = 0.65)) +
  scale_color_brewer(palette = "Set2") +
  facet_grid(~ order)

}
\seealso{
Other blending functions and operators: 
\code{\link{blend}},
\code{\link{pow-ggblend}},
\code{\link{times-ggblend}}
}
\concept{blending functions and operators}
